requirements.txt
flask==2.1.2

psycopg2-binary==2.8.6